Flyttet kode for blankett fra routes til en egen fil. Endret dependencies.
