+++ package.json
@@ -32,9 +32,7 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
-    "nodemon": "^3.0.1",
     "react": "^18.2.0",
-    "react-scripts": "^5.0.1",
     "date-fns": "^2.30.0"
   },
   "devDependencies": {
@@ -62,7 +60,8 @@
     "ts-loader": "^9.4.4",
     "webpack-cli": "^5.1.4",
     "webpack-node-externals": "^3.0.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1"
   },
   "browserslist": {
     "production": [
